Update GridComplexExample.js
When the user is done with the registration they click on the submit button and then the page should switch to the restaurant page. The comments that I added to the file seem to be the answer but I haven't tested it on the actual program yet and that is why it is commented.